Code changes + test
    Added test for sieve and did some
    modification to algorithm because of
    bugs/defects found during testing.

+++ inc/misc.h
@@ -7,7 +7,7 @@
 // definitions in misc.cpp
 namespace mz {
 
-std::vector<int> eratosthenes_sieve(size_t limit);
+std::vector<int> eratosthenes_sieve(short limit);
 
 int greatest_common_divisor(int a, int b);
